Services and handlers
Adding extra state for cancellation is kind of a PITA but cancellation just waits for the export to finish without it.
This is mixing exceptions and empty strings--I'm coming from a C++ background and exceptions kind of scare me, so I tried to avoid them unless it looked like a big diff to avoid them the NoRecordingInProgressErrors had a precedent so I got a bit lazy.