Implement bulk export (#770)
* New UI

Add a button to the side panel overlay for bulk export
and an overlay for the export in progress

* Services and handlers

Adding extra state for cancellation is kind of a PITA but cancellation just waits for the export to finish without it.
This is mixing exceptions and empty strings--I'm coming from a C++ background and exceptions kind of scare me, so I tried to avoid them unless it looked like a big diff to avoid them the NoRecordingInProgressErrors had a precedent so I got a bit lazy.

* Piping and logic for card creation and export

Like the last change, this involves some error softening as we plow through an export

* The whole export loop in SidePanel.tsx

* First pass at moving logic to the controller

This still tests alright

* Remove the rest of the changes from mining-context

* Push exportMode from CardModel to Message

This state was already in a few other messages, so we just had to add it to CardExportedMessage().

* Replace empty string error handling with exceptions

Adds DuplicateNoteError for when a card already exists
Adds AudioRequestSupersededError for when the recorder gets preempted

* Not from code review: cut down on the `as any`

* Not from code review: notify user on duplicate card

Before this, we'd just display "card exported: x/y" as before
I'm also removing the initial notify progress, which would display "card exported: x/y" before we even started exporting.

* Not from code review: disable bulk export while a normal export is happening

This puts the app in a weird state so this just greys out the control in the overlay while a card is recording

* Remove AnkiExportMode.bulk enum and refactor to use separate methods

Replace the 'bulk' export mode with a boolean flag and separate publish
methods to simplify the branching logic in CardPublisher.
Addresses:
"Sorry for realizing this so late, but it doesn't seem like a distinct..."
"Is there a reason you're writing this instead of..."
"I would recommend writing a new public method specifically for bulk export..."
"exportMode appears to be unused..."
"As mentioned in my comment on model.ts..."
"As mentioned elsewhere, I think you should branch here..."

* Just make bulkExportCancelled a public bool

Addresses "Nitpick: it's probably more efficient..."

* Fire-and-forget in bulk export controller

Addresses "I think in this case fire-and-forget is fine..."

* Also notify progress on generic errors

Addresses "Also: if CardExportedMessage.exportError is..."

* Remove AudioRequestSupersededError

Responds to:
"In this case, would we be exporting a card..."
"Was this change necessary? From what I can tell, re-record..."

* Remove dead forced UI update

* Move duplicate error throw into general code; remove extra string check

Addresses "This if -> throw block can be moved into _executeAction" and my response to it

* Run yarn pretty

* Remove erroring line

* Stop dropping exceptions on the floor and remove duplicate sendMessage

* Use BulkExportStartedPayload

* Revert control flow changes in drm error catch

* Return with index in currentSubtitle

Code review aside, we also get to remove some logic in the BulkExportController

* Fix build errors

---------

Co-authored-by: R-J Lim <kgerbil@gmail.com>